Relocate Chunk cursor
